requirements.txt
discord.py
requests
webcolors
async_cleverbot
git+https://github.com/Rapptz/discord-ext-menus
discord-py-slash-command
discord-components
